+++ tests/system/pcapfs-system-tests.py
@@ -49,6 +49,8 @@ class TestXor:
         with mount_pcap(test_pcap, params=['-c', '{here}/configs/xor.toml'.format(here=HERE),
                                            '-k', '{here}/keyfiles/xor.key'.format(here=HERE)]) as mountpoint:
             assert get_file_list(mountpoint) == expected_files_with_xor
+            with open(os.path.join(mountpoint, 'xor/10-0_xor')) as f:
+                assert f.read() == 'pcapFStest'
 
 
 class TestSsl: